refactor and move the functionality of downloading images to more of a scripts area since it'll be infrequent/not need for API functionality
